Warning inferno - fix warnings, then fix warnings

Closes #8674

See merge request OpenMW/openmw!4927

+++ apps/openmw/mwgui/quickkeysmenu.hpp
@@ -101,7 +101,7 @@ namespace MWGui
         QuickKeysMenu* mParent;
 
         bool onControllerButtonEvent(const SDL_ControllerButtonEvent& arg) override;
-        int mControllerFocus = 0;
+        size_t mControllerFocus = 0;
     };
 
     class MagicSelectionDialog : public WindowModal
@@ -122,7 +122,7 @@ namespace MWGui
         void onModelIndexSelected(SpellModel::ModelIndex index);
 
         bool onControllerButtonEvent(const SDL_ControllerButtonEvent& arg) override;
-        int mControllerFocus = 0;
+        size_t mControllerFocus = 0;
     };
 }
